rename api packages for clarity
